Implement credentials data source.
We've seen issues where the entire value of a set of credentials for a
resource/project gets cached and only being refreshed on `apply`.

When a user wants to use a new credential, this can cause issues. When
running `plan`, the value isn't available yet but the user might want to
reference this already. It only becomes available on `apply` when the
structure is changed, which it doesn't until we reference the new key.
Which in turn we can't unless it's being refreshed.

By referencing key per key, we introduce a mechanic that allows people
to not run into this issue. It does however means that users have to add
extra credentials.